bug 1542390: rewrite bugzilla associations crontabber job
This rewrites the bugzilla association crontabber job as a Django command.
This also adds some needed features in cronrun--it passes the last_success
time now.
