+++ socorro/cron/crontabber_app.py
@@ -1089,7 +1089,6 @@ DEFAULT_JOBS = ','.join([
     'socorro.cron.jobs.archivescraper.ArchiveScraperCronApp|1h',
 
     # Crash data analysis
-    'socorro.cron.jobs.bugzilla.BugzillaCronApp|1h',
     'socorro.cron.jobs.update_signatures.UpdateSignaturesCronApp|1h',
 
     # Dependency checking
