+++ socorro/unittest/cron/test_crontabber_app.py
@@ -8,7 +8,6 @@ from socorro.cron.crontabber_app import jobs_converter, DEFAULT_JOBS
 class TestJobsConverter(object):
     expected_job_names = [
         'ArchiveScraperCronApp',
-        'BugzillaCronApp',
         'DependencySecurityCheckCronApp',
         'ElasticsearchCleanupCronApp',
         'UpdateSignaturesCronApp',
